Add MathB 0.1
The source code contains the MathB and Susam libraries that
provide the classes that can be used to implement the
mathematics pastebin web application. Additionally, it provides
two implementations of the web application:

  1. main.php    -- An implementation of the application with
                    the default view
  2. mathbin.php -- An implementation of the application to be
                    hosted on http://mathb.in/

This source tree does not contain third-party JavaScript
libraries such as MathJax and Markdown which are essential for
the web application to function correctly. They need to be
installed separately. The steps to do so are included in
the README.md file.