Reduce height of Facebook like box to 188 pixels
The Facebook like box is too long vertically. Its default height
is 241 pixels. Reduce its height to 188 pixels so that it can
display only one row of profile pictures. Detailed calculation
of this new height is provided below.

The minimum height required to contain two rows of profile
pictures in the Facebook widget is the sum of the following
heights.

  1. Height of the header: 10 + 49 + 10 = 69
  2. First divider border: 1
  3. Height of the like count element: 5 + 14 = 19
  4. Height of the first row of profile pictures: 3 + 50 = 53
  5. Height of the second row of profile pictures: 3 + 50 = 53
  6. Padding around the list of profile pictures: 10 + 10 = 20
  7. Second divider border: 1
  8. Height of the footer: 5 + 10 + 5 = 20
  9. Border around the widget: 1 + 1 = 2

Therefore, the minimum height required to contain two rows of
profile pictures is:

    69 + 1 + 19 + 53 + 53 + 20 + 1 + 20 + 2 = 238

However, the actual height of the default Facebook widget has
been found to be 241 pixels, i.e. 3 pixels more than the minimum
required height.

Similar calculation shows that the widget must be at least
185 pixels high in order to contain one row of profile pictures.

    69 + 1 + 19 + 53 + 20 + 1 + 20 + 2 = 185

After adding an extra 3 pixels, the height turns out to be
188 pixels.