Fix: Path of the log file is prepended to each log
Due to a bug in Susam\Session, the path of the log file is being
included after every line of log. Further, this line containing
the path of the log file is not terminated with a newline.
Therefore, it becomes a prefix of the next log. As a result,
every line of log is prepended with the path of the log file.
Fix this issue.