Add IP address blacklisting
Add the capability to blacklist IP addresses. This is useful to
prevent spammers from submitting posts. The MathB\Configuration
object now contains a property called ipBlacklist which may be
used by an application to specify an array of regular
expressions. If the client's IP address matches one of these
regular expressions, it won't be allowed to submit posts.