Parse Markdown code before sanitizing HTML
The PHP code that runs at server-side sanitizes the HTML before
parsing the Markdown code. This is an issue because this causes
the sanitizer to mistake text or URLs within angle brackets for
HTML tags, and remove them. Examples of such text are
<http://mathb.in>, <stdio.h> in a code block, etc. We want
Markdown parser to first process the text so that it properly
converts such text into proper HTML entities or tags before the
sanitizer kicks in.