Show static preview when JavaScript is disabled
The static preview is displayed as a PNG image generated from
the input code. This PNG image is generated using pandoc and
convert. The image is created on the first requested to the PNG
preview image, and then cached in a cache directory. For
subsequent requests, the preview image is picked up from this
cache.