Display preview button when JavaScript is disabled
When JavaScript is disabled, the user cannot see a live preview
of the rendered code. A preview button can help the user to see
a static preview (rendered with an image) before he or she saves
the post. The preview button should be displayed if and only if
JavaScript is disabled.