Remove lib/mathb directory from include path
The lib directory contains the mathb package. The lib/mathb
directory contains classes. For autoloading classes using
spl_autoload, the include path should contain the path to the
directory that contains the packages (directories corresponding
to namespaces), not the directory that contains the classes.

Discussion about why lib/mathb was added to the include path and
why it is being removed is present at
<https://github.com/susam/mathb/pull/1>.