Remove rewrite rule for PNG preview from .htaccess
The .htaccess file has an additional rule to map requests for
PNG preview files to the application PHP script. Eliminate this
rule. Instead, let the URLs to request PNG preview files be the
application home page URL with the hash of the post preview as a
query parameter appended to the home page URL.