Display newlines in error messages in the preview
When an error occurs while converting a post to PNG preview,
the newlines in the error messages are displayed as \n in the
error message. As a result, the newlines are missing from the
error message displayed in the PNG image.

To reproduce the issue, disable JavaScript, create a new post
with the the following content only.

    \foo

Hit the preview button. The following error message would be
displayed.

    pandoc: Error producing PDF from TeX source.\n! Undefined
    control sequence.\nl.46 \foo\n

We want to display the following instead.

    pandoc: Error producing PDF from TeX source.
    ! Undefined control sequence.
    l.46 \foo

Fix this error by joining the output/error lines obtained from
the execution of the pandoc/convert command with newline
character instead of the two byte string '\n'.