Make static preview an optional feature
There is no way to opt out of the static preview feature. It
means that the administrator hosting this application either
needs to install the packages required for static preview, or
end up with error logs in the Apache error log file. When a user
uses the application from a browser with JavaScript disabled, if
the application doesn't succeed in creating a static preview
with the help of packages like pandoc, texlive and imagemagick,
then it would log errors in the Apache error log file.

Make this feature optional, so that if these packages are not
installed, the administrator doesn't have to deal with errors
related to static preview in the logs. This feature would now be
disabled by default, and enabled if required by the
administrator.

Update README.md with information about static preview and the
steps to enable static preview.