Add some margin between output code and date
When the output sheet contains rendered code that is long enough
to stretch the end close to the date at the bottom right corner
of the output sheet, the rendered code and the date appear to be
too close to each other. Add some separation between the
rendered code and the date by adding top margin to the date.