Do not update output sheet when input loses focus
When an input field loses focus, the output sheet is updated.
This causes the HTML in the output sheet to be rendered again
despite no change. Since the output sheet is updated on keyup,
paste and cut events, the output sheet remains updated with the
most recent input just before the input field loses focus.
Therefore, there is no need to update the output sheet again
just after the input field loses focus.