Fix rendering error for input requiring extensions
MathB.in loads the MathJax combined component named tex-mml-chtml. The
documentation for this component in the MathJax v3.2 documentation
states:

> The input/tex component loads the ams, newcommand, require,
> autoload, configmacros, and noundefined extensions, which means that
> most other extensions will be loaded automatically when needed, or
> you can use the `\require` macro to load them explicitly.

See https://docs.mathjax.org/en/v3.2-latest/web/components/combined.html
for the above documentation.

Let us consider a LaTeX input like `\bcancel{10}` which requires the
an extension named `cancel` to be autoloaded. When we type such an
input in the textarea field of MathB, we expect MathJax to autoload
the necessary extension and render the LaTeX input successfuly.

However, prior to this fix, typing such a LaTeX input leads to the
following error:

    Uncaught Error: MathJax retry

This error occurs because MathB calls `MathJax.typeset()` to typeset
the page. This function is documented to throw an error when the input
contains LaTeX commands that requires extensions to be autoloaded. The
documentation states:

> The first of these, `MathJax.typeset()`, typesets the page, and does
> so immediately and synchronously, so when the call finishes, the
> page will have been typeset. Note, however, that if the math
> includes actions that require additional files to be loaded (e.g.,
> TeX input that uses *require*, or that includes autoloaded
> extensions), then an error will be thrown. You can use the
> `try/catch` command to trap this condition.
>
> The second, `MathJax.typesetPromise()`, performs the typesetting
> asynchronously, and returns a promise that is resolved when the
> typesetting is complete. This properly handles loading of external
> files, so if you are expecting to process TeX input that can include
> `require` or autoloaded extensions, you should use this form of
> typesetting. It can be used with `await` as part of a larger `async`
> function.

See https://docs.mathjax.org/en/v3.2-latest/web/typeset.html for the
above documentation.

This change replaces `MathJax.typeset()` with
`MathJax.typesetPromise()` so that rendering completes successfully
even when the LaTeX input requires autoloading of MathJax extensions.