Fix missed sanitization of alternate HTML elements
Prior to this fix, alternate HTML elements are not being sanitized.
For example, the `<b>` tag is not an allowed tag in user posts.
However, when the following input is entered in the code textarea:

    <b>foo</b><b>bar</b><b>baz</b><b>qux</b>

the following output is produced:

    barqux

The alternate `<b>` elements for `<b>bar</b>` and `<b>qux</b>` are not
filtered out due to sanitization being missed for alternate elements.
This issue occurs due to a faulty loop that traverses the list of
child nodes of an HTML element in forward order which causes the
position of child node elements to shift as they are deleted from the
list.

This fix resolves the issue by traversing the list of child nodes in
reverse order so that the position of child nodes not visited yet do
not shift as the visited child nodes are deleted.